quick_cache

Module rw_lock

Structs

RwLock

A reader-writer lock.

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.

Type Aliases

InnerRwLock 🔒

InnerRwLockReadGuard 🔒

InnerRwLockWriteGuard 🔒