quick_cache

Module shard

Macros

record_hit 🔒

record_hit_mut 🔒

record_miss 🔒

record_miss_mut 🔒

Structs

CacheShard

A bounded cache using a modified CLOCK-PRO eviction policy. The implementation allows some parallelism as gets don’t require exclusive access. Any evicted items are returned so they can be dropped by the caller, outside the locks.

Placeholder 🔒

RefMut

Structure wrapping a mutable reference to a cached item.

Resident

Enums

Entry 🔒

InsertStrategy

ResidentState 🔒

Constants

MAX_F 🔒

Traits

SharedPlaceholder